% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_CogState.R
\name{BHR_CogState}
\alias{BHR_CogState}
\title{\strong{BHR CogState Module}}
\usage{
BHR_CogState(DET, IDN, OCL, ONB, TimePoint = "ALL_COG", Return = "SPEED_ACC")
}
\arguments{
\item{DET}{Raw dataframe extracted from the detection (DET) task in the Brain Health Registry CogState Module}

\item{IDN}{Raw dataframe extracted from the identification (IDN) task in the Brain Health Registry CogState Module}

\item{OCL}{Raw dataframe extracted from the one card learning (OCL) task in the Brain Health Registry CogState Module}

\item{ONB}{Raw dataframe extracted from the one back (ONB) task in the Brain Health Registry CogState Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a complete CogState assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a complete CogState assessment.
(3) "ALL_COG": For each participant, return all time points with a complete CogState assessment.
(4) "ALL": For each participant, return all time points with CogState data, regardless of whether the assessment is complete (i.e., data may be missing for some variables)}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "SPEED_ACC": Output dataframe includes measures of accuracy and speed for all 4 CogState tasks
(2) "ALL": Output dataframe includes all CogState measures}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.CogState:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.CogState:} {See BHR Codebook (StatusDateTime)}
 \item \code{Valid_Test:} {Identifies valid tests passing the BHR integrity check in each CogState domain. Coded as Invalid=0, Valid=1}
 \item \code{dur.DET:} {See BHR Codebook (Detection Task)}
 \item \code{lmn.DET:} {See BHR Codebook (Detection Task)}
 \item \code{lsd.DET:} {See BHR Codebook (Detection Task)}
 \item \code{acc.DET:} {See BHR Codebook (Detection Task)}
 \item \code{cor.DET:} {See BHR Codebook (Detection Task)}
 \item \code{err.DET:} {See BHR Codebook (Detection Task)}
 \item \code{presnt.DET:} {See BHR Codebook (Detection Task)}
 \item \code{sti.DET:} {See BHR Codebook (Detection Task)}
 \item \code{WasInterrupted.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in September 2014)}
 \item \code{TestingEnvironment.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in September 2014)}
 \item \code{SelfAssessment.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in Novemver 2016)}
 \item \code{GoodMeasure.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in Novemver 2016)}
 \item \code{Experience.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in Novemver 2016)}
 \item \code{InstructionsClear.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in Novemver 2016)}
 \item \code{ExplanationHelpful.DET:} {See BHR Codebook (Detection Task; Added to CogState Module in Novemver 2016)}
 \item \code{IntegrityCheck.DET:} {See BHR Codebook (Detection Task)}
 \item \code{dur.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{lmn.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{lsd.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{acc.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{cor.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{err.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{presnt.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{sti.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{WasInterrupted.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in September 2014)}
 \item \code{TestingEnvironment.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in September 2014)}
 \item \code{SelfAssessment.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in Novemver 2016)}
 \item \code{GoodMeasure.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in Novemver 2016)}
 \item \code{Experience.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in Novemver 2016)}
 \item \code{InstructionsClear.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in Novemver 2016)}
 \item \code{ExplanationHelpful.IDN:} {See BHR Codebook (Identification Task; Added to CogState Module in Novemver 2016)}
 \item \code{IntegrityCheck.IDN:} {See BHR Codebook (Identification Task)}
 \item \code{dur.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{lmn.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{lsd.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{acc.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{cor.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{err.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{presnt.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{sti.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{WasInterrupted.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in September 2014)}
 \item \code{TestingEnvironment.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in September 2014)}
 \item \code{SelfAssessment.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in Novemver 2016)}
 \item \code{GoodMeasure.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in Novemver 2016)}
 \item \code{Experience.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in Novemver 2016)}
 \item \code{InstructionsClear.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in Novemver 2016)}
 \item \code{ExplanationHelpful.OCL:} {See BHR Codebook (One Card Learning Task; Added to CogState Module in Novemver 2016)}
 \item \code{IntegrityCheck.OCL:} {See BHR Codebook (One Card Learning Task)}
 \item \code{dur.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{lmn.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{lsd.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{acc.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{cor.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{err.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{presnt.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{sti.ONB:} {See BHR Codebook (One Back Task)}
 \item \code{WasInterrupted.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in September 2014)}
 \item \code{TestingEnvironment.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in September 2014)}
 \item \code{SelfAssessment.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in Novemver 2016)}
 \item \code{GoodMeasure.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in Novemver 2016)}
 \item \code{Experience.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in Novemver 2016)}
 \item \code{InstructionsClear.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in Novemver 2016)}
 \item \code{ExplanationHelpful.ONB:} {See BHR Codebook (One Back Task; Added to CogState Module in Novemver 2016)}
 \item \code{IntegrityCheck.ONB:} {See BHR Codebook (One Back Task)}
}
}
\description{
This function cleanses data from the Brain Health Registry CogState Module. The user is asked to input four raw dataframes from the BHR CogState Module. Dataframes include (1) CogState DET (i.e., the detection test), (2) CogState IDN (i.e., the identification test), (3) CogState OCL (i.e., the one card learning test), and (4) CogState ONB (i.e., the one back test). Additionally, the user is asked to specify to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point,  last timepoint, etc.). The function performs basic data cleaning to merge the four dataframes, remove duplicate observations (including obsertvations less than 30 days from the previous time point) and assess the validity of each observation/assessment using pre-coded integrity checks.
}
