% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Demographics.R
\name{BHR_Demographics}
\alias{BHR_Demographics}
\title{\strong{BHR Demograhics Module}}
\usage{
BHR_Demographics(dataset, TimePoint = "ALL_DEMO")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Demographics Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with complete BMI information.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a complete BMI information.
(3) "ALL_DEMO": For each participant, return all time points with complete BMI information.
(4) "ALL": For each participant, return all time points with demographics data, regardless of whether the demographics assessment is complete (i.e., data may be missing for some variables)}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.Demo:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.Demo:} {See BHR Codebook (StatusDateTime)}
 \item \code{QID177:} {See BHR Codebook}
 \item \code{QID178:} {See BHR Codebook}
 \item \code{BMI_6cat:} {Body Mass Index (6 categories). Coded as 1=Underweight, 2=Normal Weight, 3=Overweight, 4=Obeseity Class I, 5=Obesity Class II, 6=Obesity Class III}
 \item \code{BMI_3cat:} {Body Mass Index (3 categories). Coded as 1=Underweight/Normal Weight, 3=Overweight, 4=Obeseity}
 \item \code{QID180:} {See BHR Codebook}
 \item \code{QID189:} {See BHR Codebook}
 \item \code{QID182:} {See BHR Codebook}
 \item \code{QID183.TEXT:} {See BHR Codebook}
 \item \code{QID185:} {See BHR Codebook}
 \item \code{QID186:} {Year of retirement, Recoded with Not applicable (i.e., not retired)=99}
 \item \code{QID192:} {See BHR Codebook}
 \item \code{QID193:} {Please indicate which branch of the Armed Forces, Recoded with Not applicable (i.e., not a veteran of the Armed Forces)=99}
 \item \code{QID184:} {See BHR Codebook}
 \item \code{QID197:} {See BHR Codebook (Added to Demographics Module in February 2021)}
 \item \code{QID199:} {See BHR Codebook (Added to Demographics Module in February 2021)}
 \item \code{QID200:} {See BHR Codebook (Added to Demographics Module in February 2021)}
 \item \code{QID201:} {See BHR Codebook (Added to Demographics Module in February 2021)}
}
}
\description{
This function cleanses data from the Brain Health Registry Demographics Module. The user is asked to input a raw dataframe from the BHR Demographics Module and to specify the desired output time point(s) from which data was collected from each participant (i.e., first time point,  last timepoint, etc.). The function  performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from  the previous time point) and to calculate participant body mass index (BMI).
}
