% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_ECog.R
\name{BHR_ECog}
\alias{BHR_ECog}
\title{\strong{BHR Everyday Cognition Module}}
\usage{
BHR_ECog(dataset, TimePoint = "ALL_ECOG", Return = "SummaryScores")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry ECog Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a completed ECog assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a completed ECog assessment.
(3) "ALL_ECOG": For each participant, return all time points with a completed ECog assessment
(4) "ALL": For each participant, return all time points with ECog data, regardless of whether the ECog assessment is complete (i.e., data may be missing for some variables)}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "SummaryScores": Output dataframe includes six domain-specific ECog scores (e.g., Memory, Language, etc.), as well as the total ECog score (mean)
(2) "ALL": Output dataframe includes all single-item and summary ECog measures.}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.ECog:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.ECog:} {See BHR Codebook (StatusDateTime)}
 \item \code{QID48} {See BHR Codebook}
 \item \code{QID49.1} {MEMORY: Remembering a few shopping items without a list. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.2} {MEMORY: Remembering things that happened recently. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.3} {MEMORY: Recalling conversations a few days later. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.4} {MEMORY: Remembering where I have placed objects. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.5} {MEMORY: Repeating stories and/or questions. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.6} {MEMORY: Remembering the current date or day of the week. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.7} {MEMORY: Remembering I have already told someone something. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID49.8} {MEMORY: Remembering appointments, meetings, or engagements. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.1} {LANGUAGE: Forgetting the names of objects. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.2} {LANGUAGE: Verbally giving instructions to others. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.3} {LANGUAGE: Finding the right words to use in a conversation. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.4} {LANGUAGE: Communicating thoughts in a conversation. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.5} {LANGUAGE: Following a story in a book or on TV. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.6} {LANGUAGE: Understanding the point of what other people are trying to say. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.7} {LANGUAGE: Remembering the meaning of common words. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.8} {LANGUAGE: Describing a program I have watched on TV. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID50.9} {LANGUAGE: Understanding spoken directions or instructions. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.1} {VISUAL: Following a map to find a new location. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.2} {VISUAL: Reading a map and helping with directions when someone else is driving. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.3} {VISUAL: Finding my car in a parking lot. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.4} {VISUAL: Finding my way back to a meeting spot in the mall or other location. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consist Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.6} {VISUAL:  Finding my way around a familiar store. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID51.7} {VISUAL: Finding my way around a house visited many times. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID52.1} {PLANNING: Planning a sequence of stops on a shopping trip. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID52.2} {PLANNING: The ability to anticipate weather changes and plan accordingly. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID52.3} {PLANNING: Developing a schedule in advance of anticipated events. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID52.4} {PLANNING: Thinking things through before acting. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID52.5} {PLANNING: Thinking ahead. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.1} {ORGANIZATION: Keeping living and work space organized. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.2} {ORGANIZATION: Balancing the checkbook without error. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.3} {ORGANIZATION: Keeping financial records organized. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.4} {ORGANIZATION: Prioritizing tasks by importance. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.5} {ORGANIZATION: Keeping mail and papers organized. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID53.6} {ORGANIZATION: Using an organized strategy to manage a medication schedule. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID54.1} {ATTENTION: The ability to do two things at once. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID54.2} {ATTENTION: Returning to a task after being interrupted. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID54.3} {ATTENTION: The ability to concentrate on a task without being distracted. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
 \item \code{QID54.4} {ATTENTION: Cooking or working and talking at the same time. Recoded as: Better or No Change=1, Questionable/Occasionally Worse=2, Consistently a Little Worse=3, Consistently Much Worse = 4, I don't know=NA}
}
}
\description{
This function cleanses data from the Brain Health Registry Everyday Cognition (ECog) Module. The user is asked to input a raw dataframe from the BHR ECog Module and to specify to specify the desired output time point(s) from which data was collected from each participant  (i.e., first time point,  last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and recode single items measures of cognitive performance. Standard scoring algorithms are used to calculate total scores on ECog subscales, as well as a ECog total/global score.
}
