% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Hoarding.R
\name{BHR_Hoarding}
\alias{BHR_Hoarding}
\title{\strong{BHR Hoarding Module}}
\usage{
BHR_Hoarding(dataset, TimePoint = "ALL_HRS", Return = "HRS")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Hoarding Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a completed HRS-SR assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a completed HRS-SR assessment.
(3) "ALL_HRS": For each participant, return all time points with a completed HRS-SR assessment.
(4) "ALL": For each participant, return all time points with hoarding data, regardless of whether the HRS-SR was completed at the given time point.}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "HRS": Output dataframe includes variables HRS-SR single-item and total scores only}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.Hoarding:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.Hoarding:} {See BHR Codebook (StatusDateTime)}
 \item \code{HRS_Clutter:} {Because of the clutter or number of possessions, how difficult is it for you to use the rooms in your home? Recoded using a scale of 0 to 8, with higher scores indicating greater severity}
 \item \code{HRS_Discard:} {To what extent do you have difficulty discarding (or recycling, selling, giving away) ordinary things that other people would get rid of? Recoded using a scale of 0 to 8, with higher scores indicating greater severity}
 \item \code{HRS_Acquire:} {To what extent do you currently have a problem with collecting free things or buying more things than you need or can use or can afford? Recoded using a scale of 0 to 8, with higher scores indicating greater severity}
 \item \code{HRS_Distress:} {To what extent do you experience emotional distress because of clutter, difficulty discarding or problems with buying or acquiring things? Recoded using a scale of 0 to 8, with higher scores indicating greater severity}
 \item \code{HRS_Impairment:} {To what extent do you experience impairment in your life (daily routine, job/school, social activities, family activities, financial difficulties) because of clutter, difficulty discarding, or problems with buying or acquiring things? Recoded using a scale of 0 to 8, with higher scores indicating greater severity}
 \item \code{HRS_Total:} {Total score on the Hoarding Rating Scale, Self-Report. Scores range from 0 to 40, with high schores indicating greater hoarding symtomatology. In the BHR, cutoffs of 10 and 14 have been used for identifying subsclinical and clinically relevant hoarding, respectively}
 \item \code{HRS_Class:} {Hoarding classification using HRS total scores. Coded as 0=No Hoarding (0-4), 1=Minimal Hoarding (5-9), 2=Subclinical Hoarding (10-13), 3=Clinically Relevant Hoarding (14-40)}
 \item \code{QID11.1:} {ADLH item 1: Use stove. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.2:} {ADLH item 2: Use kitchen counters. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.3:} {ADLH item 3: Eat at table. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.4:} {ADLH item 4: Use bath/shower. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.5:} {ADLH item 5: Sit in sofa/chair. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.6:} {ADLH item 6: Sleep in bed. Scores of 6 (NA) are recoded as NA}
 \item \code{QID11.7:} {ADLH item 7: Find important things. Scores of 6 (NA) are recoded as NA}
 \item \code{QID14.1:} {See BHR Codebook}
 \item \code{QID14.2:} {See BHR Codebook}
 \item \code{QID14.3:} {See BHR Codebook}
 \item \code{QID14.4:} {See BHR Codebook}
 \item \code{QID14.5:} {See BHR Codebook}
 \item \code{QID44.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID44.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID45.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID45.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID46.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID46.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID47.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID47.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID48.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID48.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID49.1:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{QID49.2:} {See BHR Codebook (Added to Hoarding Module in February 2019)}
 \item \code{WHODAS_Total:} {Total score on the World Health Organization Disability Asssessment Schedule. Scores range from 0 to 48, with high schores indicating greater disability.}
 \item \code{QID49.4:} {See BHR Codebook}
 \item \code{QID50:} {See BHR Codebook}
 \item \code{QID51:} {See BHR Codebook}
 \item \code{QID52:} {See BHR Codebook}
 \item \code{QID39:} {See BHR Codebook}
 \item \code{QID32:} {See BHR Codebook}
 \item \code{QID33.1:} {See BHR Codebook}
 \item \code{QID33.2:} {See BHR Codebook}
 \item \code{QID33.3:} {See BHR Codebook}
 \item \code{QID33.4:} {See BHR Codebook}
 \item \code{QID33.5:} {See BHR Codebook}
 \item \code{QID33.6:} {See BHR Codebook}
 \item \code{QID33.7:} {See BHR Codebook}
 \item \code{QID33.7.TEXT:} {See BHR Codebook}
 \item \code{QID34:} {Do you breed animals? Recoded as Yes=1, No=2, Not applicable (i.e., no pets)=99 }
 \item \code{QID35:} {Has anyone ever told you that you have too many pets? Recoded as Yes=1, No=2, Not applicable (i.e., no pets)=99}
 \item \code{QID36:} {Have you ever had difficulty in caring for your animals for financial or other reasons? Recoded as Yes=1, No=2, Not applicable (i.e., no pets)=99}
 \item \code{QID37:} {Has anyone ever told you that they were worried about your ability to care for your animals? Recoded as Yes=1, No=2, Not applicable (i.e., no pets)=99}
 \item \code{QID15.TEXT:} {See BHR Codebook}
 \item \code{QID16.TEXT:} {See BHR Codebook}
 \item \code{QID20:} {See BHR Codebook}
 \item \code{QID21:} {See BHR Codebook}
 \item \code{QID22:} {See BHR Codebook (Removed from Hoarding Module in February 2019)}
 \item \code{QID23:} {See BHR Codebook (Removed from Hoarding Module in February 2019)}
 \item \code{QID24:} {See BHR Codebook (Removed from Hoarding Module in February 2019)}
 \item \code{QID25:} {See BHR Codebook (Removed from Hoarding Module in February 2019)}
 \item \code{QID26.1:} {See BHR Codebook}
 \item \code{QID26.2:} {See BHR Codebook}
 \item \code{QID26.3:} {See BHR Codebook}
 \item \code{QID26.4:} {See BHR Codebook}
 \item \code{QID26.5:} {See BHR Codebook}
 \item \code{QID26.6:} {See BHR Codebook}
 \item \code{QID26.7:} {See BHR Codebook}
 \item \code{QID26.8:} {See BHR Codebook}
 \item \code{QID27:} {See BHR Codebook}
 \item \code{QID41:} {See BHR Codebook}
 \item \code{QID28:} {See BHR Codebook}
 \item \code{QID29:} {See BHR Codebook}
 \item \code{QID42:} {See BHR Codebook}
}
}
\description{
This function cleanses data from the Brain Health Registry Hoarding Module. The user is asked to input a raw dataframe from the BHR Hoarding Module and to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point, last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and to recode single items measures of hoarding symptoms and disability on the Hoarding Rating Scale, Self-Report (HRS-SR) and the Activities of Daily Living in Hoarding (ADLH). Standard scoring algorithms are used to calculate total scores on the HRS-SR and the World Health Organization Disability Asssessment Schedule (WHODAS).
}
