% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Medical.R
\name{BHR_Medical}
\alias{BHR_Medical}
\title{\strong{BHR Medical Module}}
\usage{
BHR_Medical(dataset, TimePoint = "ALL_MED", Return = "ALL_MED")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Medical Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a completed medical assessment (i.e., medical, neuro, and psychiatric health conditions assessed).
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a completed medical assessment.
(3) "ALL_MED": For each participant, return all time points with a completed medical assessment.
(4) "ALL": For each participant, return all time points with medical data, regardless of whether the medical assessment is complete (i.e., data may be missing for some health conditions)}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "MED_NEURO": Output dataframe includes medical and nuerological health conditions only
(2) "PSYCH": Output dataframe includes psychiatric health conditions (and lifetime psychiatric variables) only
(3) "ALL_MED": Output dataframe includes all medical variables, including lifetime psychiatric health history variables and variables measuring disease burden
(3) "ALL": All data columns, including medical, psychiatric, and neurological health conditions, as well as information regarding surgeries}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
#' \itemize{
 \item \code{SubjectCode:} {See BHR Codebook}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.Med:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.Med:} {See BHR Codebook (StatusDateTime)}
 \item \code{QID1.3} {Lifetime history of stroke. Recoded as No=0, Yes=1}
 \item \code{QID1.6} {Lifetime history of heart disease. Recoded as No=0, Yes=1}
 \item \code{QID1.7} {Lifetime history of high blood pressure. Recoded as No=0, Yes=1}
 \item \code{QID1.8} {Lifetime history of high cholesterol. Recoded as No=0, Yes=1}
 \item \code{QID1.9} {Lifetime history of diabetes. Recoded as No=0, Yes=1}
 \item \code{QID1.10} {Lifetime history of cancer. Recoded as No=0, Yes=1}
 \item \code{QID1.15} {Lifetime history of lung disease. Recoded as No=0, Yes=1  (Added to Medical Module in April 2015)}
 \item \code{QID1.16} {Lifetime history of asthma. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.20} {Lifetime history of hearing loss. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.17} {Lifetime history of arthritis. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.14} {Lifetime history of traumatic brain injury. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.18} {Lifetime history of concussion. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID2} {Lifetime history of chronic pain. Recoded as No=0, Yes=1, Prefer Not to Say=99}
 \item \code{QID3.1} {See BHR Codebook}
 \item \code{QID4} {Lifetime history of sleep apnea. Recoded as No=0, Yes=1, Prefer Not to Say=99}
 \item \code{QID6} {Lifetime history of allergies. Recoded as No=0, Yes=1, Prefer Not to Say=99}
 \item \code{Med_Burden} {Number of medical health conditions, including stroke, heart disease, high blood pressure, high cholesterol, diabetes, cancer, asthma, hearing loss, arthritis, traumatic brin injury, concussion, chronic pain, sleep apnea, and allergies. Values ranges from 0 to 16, with higher values indicating a greater number of medical conditions}
 \item \code{QID1.1} {Lifetime history of Parkinson's disease. Recoded as No=0, Yes=1}
 \item \code{QID1.2} {Lifetime history of movement disorder. Recoded as No=0, Yes=1}
 \item \code{QID1.24} {Lifetime history of essential tremor. Recoded as No=0, Yes=1 (Added to Medical Module in April 2015)}
 \item \code{QID1.25} {Lifetime history of Huntington's disease. Recoded as No=0, Yes=1 (Added to Medical Module in April 2015)}
 \item \code{QID1.4} {Lifetime history of motor neuron disease. Recoded as No=0, Yes=1}
 \item \code{QID1.26} {Lifetime history of amyotrophic lateral sclerosis. Recoded as No=0, Yes=1 (Added to Medical Module in April 2015)}
 \item \code{QID1.12} {Lifetime history of Alzheimer's disease. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.13} {Lifetime history of mild cognitive impairment. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID1.5} {Lifetime history of dementia. Recoded as No=0, Yes=1}
 \item \code{QID1.22} {Lifetime history of frontotemporal dementia. Recoded as No=0, Yes=1 (Added to Medical Module in August 2014)}
 \item \code{QID1.23} {Lifetime history of Lewy body disease. Recoded as No=0, Yes=1 (Added to Medical Module in August 2014)}
 \item \code{QID1.21} {Lifetime history of multiple sclerosis. Recoded as No=0, Yes=1 (Added to Medical Module in August 2014)}
 \item \code{QID1.19} {Lifetime history of epilepsy or seizures. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{Neuro_Burden} {Number of neurological health conditions, including Parkinson's disease, movement  disorder, essential tremor, motor neuron disease, Alzheimer's disease, mild cognitive impairment, dementia, frontotemporal dementia, lewy body disease, multiple sclerosis, and epilepsy or seizures. Values ranges from 0 to 11, with higher values indicating a greater number of neurological conditions}
 \item \code{QID28.1.1} {Current major depressive disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.1} {Past major depressive disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_MDD} {Lifetime history of major depressive disorder. Coded as No=0, Yes=1}
 \item \code{MDD_Most} {Major depressive disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.3} {Current specific phobia or social phobia. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.3} {Past specific phobia or social phobia. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_Phobia} {Lifetime history of specific phobia or social phobia. Coded as No=0, Yes=1}
 \item \code{Phobia_Most} {Phobia endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.4} {Current obsessive compulsive disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.4} {Past obsessive compulsive disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_OCD} {Lifetime history of obsessive compulsive disorder. Coded as No=0, Yes=1}
 \item \code{OCD_Most} {Obsessive compulsive disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.5} {Current hoarding disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.5} {Past hoarding disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_HD} {Lifetime history of hoarding disorder. Coded as No=0, Yes=1}
 \item \code{HD_Most} {Hoarding disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.6} {Current attention deficit/hyperactivity disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.6} {Past attention deficit/hyperactivity disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_ADHD} {Lifetime history of attention deficit/hyperactivity disorder. Coded as No=0, Yes=1}
 \item \code{ADHD_Most} {Attention deficit/hyperactivity disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.8} {Current post-traumatic stress disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.8} {Past post-traumatic stress disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_PTSD} {Lifetime history of post-traumatic stress disorder. Coded as No=0, Yes=1}
 \item \code{PTSD_Most} {Post-traumatic stress disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.9} {Current generalized anxiety disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.9} {Past generalized anxiety disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_GAD} {Lifetime history of generalized anxiety disorder. Coded as No=0, Yes=1}
 \item \code{GAD_Most} {Generalized anxiety disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.10} {Current panic disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.10} {Past panic disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_Panic} {Lifetime history of panic disorder. Coded as No=0, Yes=1}
 \item \code{Panic_Most} {Panic disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.11} {Current bipolar disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.11} {Past bipolar disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_BPD} {Lifetime history of bipolar disorder. Coded as No=0, Yes=1}
 \item \code{BPD_Most} {Bipolar disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.12} {Current autism spectrum disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.12} {Past autism spectrum disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_ASM} {Lifetime history of autism spectrum disorder. Coded as No=0, Yes=1}
 \item \code{ASM_Most} {Autism spectrum disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.13} {Current schizophrenia. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.13} {Past schizophrenia. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_SCZ} {Lifetime history of schizophrenia. Coded as No=0, Yes=1}
 \item \code{SCZ_Most} {Schizophrenia endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.14} {Current eating disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.14} {Past eating disorder. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_ED} {Lifetime history of eating disorder. Coded as No=0, Yes=1}
 \item \code{ED_Most} {Eating disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID28.1.15} {Current psychosis. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{QID28.2.15} {Past psychosis. Recoded as No=0, Yes=1 (Added to Medical Module in July 2014)}
 \item \code{LT_Psychosis} {Lifetime history of psychosis. Coded as No=0, Yes=1}
 \item \code{Psychosis_Most} {Psychosis endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{QID9} {Lifetime history of alcohol abuse. Recoded as No=0, Yes=1, Prefer Not to Say = 99}
 \item \code{QID13.TEXT} {See BHR Codebook}
 \item \code{QID15.TEXT} {See BHR Codebook}
 \item \code{QID16.TEXT} {See BHR Codebook}
 \item \code{QID17} {Lifetime history of drug abuse. Recoded as No=0, Yes=1, Prefer Not to Say = 99}
 \item \code{QID18.TEXT} {See BHR Codebook}
 \item \code{QID19.TEXT} {See BHR Codebook}
 \item \code{ANX} {Lifetime history of any anxiety disorders (i.e., generalized anxiety disorder, specific or social phobia, or panic disorder). Coded as No=0, Yes=1}
 \item \code{SMI} {Lifetime history of any severe mental illness (i.e., bipolar disorder, schizophrenia, or psychosis). Coded as No=0, Yes=1}
 \item \code{PsychChild} {Lifetime history of any psychiatric disorder of childhood (i.e., attention deficit/hyperactivity disorder or autism spectrum disorder). Coded as No=0, Yes=1}
 \item \code{OCRD} {Lifetime history of any obsessive compulsive related disorder (i.e., obsessive compulsive disorder or eating disorder). Coded as No=0, Yes=1}
 \item \code{SUD}  {Lifetime history of any substance use disorder (i.e., alcohol abuse or drug abuse). Coded as No=0, Yes=1}
 \item \code{SUD_Most} {Substance use disorder endorsed at most timepoints. Coded as No=0, Yes=1. NA if less than 3 timepoints completed}
 \item \code{Psych_Burden} {Number of psychiatric health conditions, including major depressive disorder, specific phobia or social phobia, obsessive compulsive disorder, hoarding disorder, attention deficit/hyperactivity disorder, post-traumatic stress disorder, generalized anxiety disorder, panic disorder, bipolar disorder, autism spectrum disorder, schizophrenia, eating disorder, psychosis, substance use (i.e., alcohol or drug abuse). Values range from 0 to 14, with higher values indicating a greater number of psychiatric conditions}
 \item \code{QID20} {Lifetime history of smoking tobacco. Recoded as No=0, Yes=1, Prefer Not to Say = 99}
 \item \code{QID21.TEXT} {See BHR Codebook}
 \item \code{QID22.TEXT} {See BHR Codebook}
 \item \code{QID23.TEXT} {See BHR Codebook}
 \item \code{QID30.1} {Current: Use of a cardiac pacemaker/defibrillator, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID30.2} {Current: Any surgical metal or any foreign objects in your body, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID30.3} {Current: Any stents, filter, or intravascular coils, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID30.4} {Current: Internal pacing wires, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID30.5} {Current: Sternum wires, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID30.6} {Current: Claustrophobia, Recoded as No/Don't know=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID31} {Work with metal (grinding, welding, etc.), Recoded as No=0, Yes=1, Prefer Not to Say = 99 (Added to Medical Module in April 2015)}
 \item \code{QID32} {Previous MRI scan, Recoded as No=0, Yes=1, Prefer Not to Say = 99  (Added to Medical Module in April 2015)}
 }
}
\description{
This function cleanses data from the Brain Health Registry Medical Module. The user is asked to input a raw dataframe from the BHR Medical Module and to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point, last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and to recode single item measures of medical, nerological, and psychiatric health history. Variables for lifetime psychiatric diagnoses are created using information on current and past diagnosis, and measures of medical, nerological, and psychiatric disease burden are calculated by summing the number of conditions endorsed in each category.
}
