% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Memtrax.R
\name{BHR_Memtrax}
\alias{BHR_Memtrax}
\title{\strong{BHR Memtrax Module}}
\usage{
BHR_Memtrax(dataset, TimePoint = "ALL_MEM", Return = "RT_ACC")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Memtrax Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a complete Memtrax assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a complete Memtrax assessment.
(3) "ALL_MEM": For each participant, return all time points with a complete Memtrax assessment.
(4) "ALL": For each participant, return all time points with Memtrax data, regardless of whether the assessment is complete (i.e., data may be missing for some variables)}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "RT_ACC": Output dataframe includes only the percentage of images answered correctly and mean reaction time in seconds, when clicking on repeat images
(2) "ALL": Output dataframe includes all Memtrax measures}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.Memtrax:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.Memtrax:} {See BHR Codebook (StatusDateTime)}
 \item \code{Valid_Test:} {Identifies valid tests with no more than 50 total images, the number of responses  between 5 and 45, and a correct response reaction time between 0.3 second and 2.3 seconds. Coded as Invalid=0, Valid=1}
 \item \code{OrderSet:} {See BHR Codebook}
 \item \code{TotalImages:} {See BHR Codebook}
 \item \code{UniqueImages:} {See BHR Codebook}
 \item \code{RepeatImages:} {See BHR Codebook}
 \item \code{CorrectN:} {See BHR Codebook}
 \item \code{CorrectPCT:} {See BHR Codebook}
 \item \code{IncorrectN:} {See BHR Codebook}
 \item \code{IncorrectPCT:} {See BHR Codebook}
 \item \code{CorrectRejectionsN:} {See BHR Codebook}
 \item \code{CorrectRejectionsPCT:} {See BHR Codebook}
 \item \code{IncorrectRejectionsN:} {See BHR Codebook}
 \item \code{IncorrectRejectionsPCT:} {See BHR Codebook}
 \item \code{TotalRejectionsN:} {See BHR Codebook}
 \item \code{TotalRejectionsPCT:} {See BHR Codebook}
 \item \code{CorrectResponsesN:} {See BHR Codebook}
 \item \code{CorrectResponsesPCT:} {See BHR Codebook}
 \item \code{CorrectResponsesRT:} {See BHR Codebook}
 \item \code{IncorrectResponsesN:} {See BHR Codebook}
 \item \code{IncorrectResponsesPCT:} {See BHR Codebook}
 \item \code{IncorrectResponsesRT:} {See BHR Codebook}
 \item \code{TotalResponsesN:} {See BHR Codebook}
 \item \code{TotalResponsesPCT:} {See BHR Codebook}
 \item \code{WasInterrupted:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{TestingEnvironment:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{SelfAssessment:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{GoodMeasure:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{Experience:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{InstructionsClear:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
 \item \code{ExplanationHelpful:} {See BHR Codebook (Added to Memtrax Module in November 2016)}
}
}
\description{
This function cleanses data from the Brain Health Registry Memtrax Module. The user is asked to input a raw dataframe from the BHR Memtrax Module and to specify to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point,  last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and assess the validity of each observation/assessment using standard criteria (i.e., no more than 50 total images per assessment, number of responses between 5 and 45, correct response reaction time between 0.3 second and 2.3 seconds).
}
