% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Prof_Particip.R
\name{BHR_Prof_Particip}
\alias{BHR_Prof_Particip}
\title{\strong{BHR Participants and Profile Modules}}
\usage{
BHR_Prof_Particip(Profile, Participants)
}
\arguments{
\item{Profile}{Raw dataframe extracted from the Brain Health Registry Participants Module}

\item{Participants}{Raw dataframe extracted from the Brain Health Registry Profile Module}
}
\value{
The function returns a dataset containing the following variables. Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{IsSubject:} {See BHR Codebook}
 \item \code{IsStudyPartner:} {See BHR Codebook}
 \item \code{BaselineDate:} {See BHR Codebook}
 \item \code{Age_Baseline:} {See BHR Codebook}
 \item \code{Gender:} {See BHR Codebook}
 \item \code{Race_Categorical:} {Race of participant. Coded as Caucasian/White=0, African American/Black=1, Asian=2, Other/Mixed=3, Declined to State=4}
 \item \code{Race_AfricanAmerican:} {See BHR Codebook}
 \item \code{Race_Asian:} {See BHR Codebook}
 \item \code{Race_Caucasian:} {See BHR Codebook}
 \item \code{Race_NativeAmerican:} {See BHR Codebook}
 \item \code{Race_PacificIslander:} {See BHR Codebook}
 \item \code{Race_Other:} {See BHR Codebook}
 \item \code{Race_DeclinedToState:} {See BHR Codebook}
 \item \code{Ethnicity:} {See BHR Codebook}
 \item \code{Edu_Categorical:} {Educational Attainment. Coded as Less than college (i.e., Grammar School, High School, Some College)=0, Two or Four Year Degree=1, Graduate/Professional Degree (i.e., Masters, Doctoral Degre, or Professional Degree)=2}
 \item \code{YearsEducationUS:} {See BHR Codebook}
 \item \code{Handedness:} {See BHR Codebook}
 \item \code{Deceased:} {See BHR Codebook}
}
}
\description{
This function cleanses data from the Brain Health Registry Participants and Profile Modules. The user is asked to input two raw dataframes from the BHR, one from the Profile Module and one from the Participants Module. The function performs basic data cleaning to merge the two dataframes and create categorical race and education variables.
}
