% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_QoL.R
\name{BHR_QoL}
\alias{BHR_QoL}
\title{\strong{BHR Quality of Life Module}}
\usage{
BHR_QoL(dataset, TimePoint = "ALL_QOL", Return = "SummaryScores")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Quality of Life Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a complete SF-36 assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a complete SF-36 assessment.
(3) "ALL_QOL": For each participant, return all time points with a complete SF-36 assessment.
(4) "ALL": For each participant, return all time points with SF-36 data, regardless of whether the assessment is complete (i.e., data may be missing for some variables)}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "SummaryScores": Output dataframe includes eight domain-specific QoL scores (e.g., physical functioning, role limitations due to physical health, etc.), as well as the two summary scores for mental and physical quality of life (i.e., mental and physical component summary scores)
(2) "SF36_Scored": Output dataframe includes eight domain-specific QoL scores, two summary scores for mental and physical QoL, and all single items measures of QoL recoded using a scale ranging from 0 to 100.
(2) "ALL": Output dataframe includes all scored and unscored SF-36 measures.}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.QoL:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.QoL:} {See BHR Codebook (StatusDateTime)}
 \item \code{QID80:} {See BHR Codebook}
 \item \code{QID81:} {See BHR Codebook}
 \item \code{QID82.1:} {See BHR Codebook}
 \item \code{QID82.2:} {See BHR Codebook}
 \item \code{QID82.3:} {See BHR Codebook}
 \item \code{QID82.4:} {See BHR Codebook}
 \item \code{QID82.5:} {See BHR Codebook}
 \item \code{QID82.6:} {See BHR Codebook}
 \item \code{QID82.7:} {See BHR Codebook}
 \item \code{QID82.8:} {See BHR Codebook}
 \item \code{QID82.9:} {See BHR Codebook}
 \item \code{QID82.10:} {See BHR Codebook}
 \item \code{QID83.1:} {See BHR Codebook}
 \item \code{QID83.2:} {See BHR Codebook}
 \item \code{QID83.3:} {See BHR Codebook}
 \item \code{QID83.4:} {See BHR Codebook}
 \item \code{QID84.1:} {See BHR Codebook}
 \item \code{QID84.2:} {See BHR Codebook}
 \item \code{QID84.3:} {See BHR Codebook}
 \item \code{QID85:} {See BHR Codebook}
 \item \code{QID86:} {See BHR Codebook}
 \item \code{QID87:} {See BHR Codebook}
 \item \code{QID140.1:} {See BHR Codebook}
 \item \code{QID140.2:} {See BHR Codebook}
 \item \code{QID140.3:} {See BHR Codebook}
 \item \code{QID140.4:} {See BHR Codebook}
 \item \code{QID140.5:} {See BHR Codebook}
 \item \code{QID140.6:} {See BHR Codebook}
 \item \code{QID140.7:} {See BHR Codebook}
 \item \code{QID140.8:} {See BHR Codebook}
 \item \code{QID140.9:} {See BHR Codebook}
 \item \code{QID88:} {See BHR Codebook}
 \item \code{QID89.1:} {See BHR Codebook}
 \item \code{QID89.2:} {See BHR Codebook}
 \item \code{QID89.3:} {See BHR Codebook}
 \item \code{QID89.4:} {See BHR Codebook}
 \item \code{Q1: In general, would you say your health is. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q2: Compared to one year ago, how would you rate your health in general now? Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q3: Does your health now limit you in these activities? If so, how much?... Vigorous activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q4: Does your health now limit you in these activities? If so, how much?... Moderate activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q5: Does your health now limit you in these activities? If so, how much?... Lifting or carrying groceries. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q6: Does your health now limit you in these activities? If so, how much?... Climbing several flights of stairs. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q7: Does your health now limit you in these activities? If so, how much?... Climbing one flight of stairs. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q8: Does your health now limit you in these activities? If so, how much?... Bending, kneeling, or stooping. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q9: Does your health now limit you in these activities? If so, how much?... Walking more than a mile. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q10: Does your health now limit you in these activities? If so, how much?... Walking several blocks. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q11: Does your health now limit you in these activities? If so, how much?... Walking one block. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q12: Does your health now limit you in these activities? If so, how much?... Bathing or dressing yourself. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q13: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of your physical health?... Cut down the amount of time you spent on work or other activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q14: PDuring the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of your physical health?... Accomplished less than you would like. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q15: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of your physical health?... Were limited in the kind of work or other activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q16: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of your physical health?... Had difficulty performing the work or other activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q17: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of any emotional problems?... Cut down the amount of time you spent on work or other activities. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q18: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of any emotional problems?... Accomplished less than you would like. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q19: During the past 4 weeks, have you had any of the following problems with your work or other regular daily activities as a result of any emotional problems?... Didn't do work or other activities as carefully as usual. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q20: During the past 4 weeks, to what extent has your physical health or emotional problems interfered with your normal social activities with family, friends, neighbors, or groups?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q21: How much bodily pain have you had during the past 4 weeks?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q22: During the past 4 weeks, how much did pain interfere with your normal work?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q23: How much of the time during the past 4 weeks...... Did you feel full of pep?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q24: How much of the time during the past 4 weeks...... Have you been a very nervous person?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q25, How much of the time during the past 4 weeks...... Have you felt so down in the dumps that nothing could cheer you up?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q26, How much of the time during the past 4 weeks...... Have you felt calm and peaceful?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q27, How much of the time during the past 4 weeks...... Did you have a lot of energy?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q28, How much of the time during the past 4 weeks...... Have you felt downhearted and blue?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q29, How much of the time during the past 4 weeks...... Did you feel worn out?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q30, How much of the time during the past 4 weeks...... Have you been a happy person?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q31, How much of the time during the past 4 weeks...... Did you feel tired?. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q32, During the past 4 weeks, how much of the time has your physical health or emotional problems interfered with your social activities? Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q33, TRUE or FALSE... I seem to get sick a little easier than other people. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q34, TRUE or FALSE... I am as healthy as anybody I know. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q35, TRUE or FALSE... I expect my health to get worse. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{Q36, TRUE or FALSE...My health is excellent. Recoded using a scale ranging from 0 to 100, where higher scores indicate better QoL}
 \item \code{PhysFunc:} {Physical Functioning Scale (mean of all items in the physical functioning domain)}
 \item \code{PhysFuncZ:} {Z-score for the Physical Functioning Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{LimitPhys:} {Role Limitations Due to Physical Health Scale (mean of all items in the role limitations domain)}
 \item \code{LimitPhysZ:} {Z-score for the Role Limitations Due to Physical Health scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{LimitEmo:} {Role Limitations Due to Emotional Health Scale (mean of all items in the role limitations domain)}
 \item \code{LimitEmoZ:} {Z-score for the Role Limitations Due to Emotional Health scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{Energy:} {Energy/Vitality Scale (mean of all items in the energy domain)}
 \item \code{EnergyZ:} {Z-score for the Energy Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{Emotional:} {Emotional Health Scale (mean of all items in the emotional health domain)}
 \item \code{EmotionalZ:} {Z-score for the Emotional Health  Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{Social:} {Social Functioning Scale (mean of all items in the social functioning domain)}
 \item \code{SocialZ:} {Z-score for the Social Functioning  Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{Pain:} {Pain Scale (mean of all items in the pain domain)}
 \item \code{PainZ:} {Z-score for the Pain  Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{Health:} {General health Scale (mean of all items in the general health domain)}
 \item \code{HealthZ:} {Z-score for the General Health  Scale. Determined by subtracting the U.S. general population scale mean from individuals' scale score and dividing by the U.S. population standard deviation.}
 \item \code{MCS_tscore:} {Mental Component Summary Score (aggregates the eight SF-36 subscales/domains using the eight z-scores and factor scoring coefficients determined by Ware et al.) MCS scores range from 0 to 100, with a mean of 50 and a standard deviation of 10 for the U.S. general population. Higher scores indicate better quality of life.}
 \item \code{PCS_tscore:} {Physical Component Summary Score (aggregates the eight SF-36 subscales/domains using the eight z-scores and factor scoring coefficients determined by Ware et al.) MCS scores range from 0 to 100, with a mean of 50 and a standard deviation of 10 for the U.S. general population. Higher scores indicate better quality of life.}
}
}
\description{
This function cleanses data from the Brain Health Registry Quality of Life (QoL) Module. The user is asked to input a raw dataframe from the BHR QoL Module and to specify to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point,  last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and recodes SF-36 (Medical Outcomes Study 36-item Short Form) single item measures using a scale ranging from 0 o 100. Eight domain-specific measures and two summary scores measuring quality of life are calculated using SF-36 recommended scoring criteria. All domain-specific and summary scores are standardized using U.S. general population means and standard deviations.
}
