% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHR_Sleep.R
\name{BHR_Sleep}
\alias{BHR_Sleep}
\title{\strong{BHR Sleep Module}}
\usage{
BHR_Sleep(dataset, TimePoint = "ALL_PSQI", Return = "PSQI")
}
\arguments{
\item{dataset}{Raw dataframe extracted from the Brain Health Registry Sleep Module}

\item{TimePoint}{Desired number of timepoints per participant to be returned in the output dataframe. Options include:
(1) "FIRST": For each participant, return the first (i.e., baseline) time point with a completed PSQI assessment.
(2) "LAST": For each participant, return the last time (i.e., most recent) point with a completed PSQI assessment.
(3) "ALL_PSQI": For each participant, return all time points with a completed PSQI assessment.
(4) "ALL": For each participant, return all time points with hoarding data, regardless of whether the PSQI was completed at the given time point.}

\item{Return}{Variables to be included in the output dataframe. Options include:
(1) "PSQI": Output dataframe includes variables PSQI component and summary scores only.
(2) "ALL": Output dataframe includes all sleep variables, including PSQI component and summary scores.}
}
\value{
The function returns a dataset containing some or all of the following variables (depending on Return parameter). Only variables recoded or created within the function are described in detail.
\itemize{
 \item \code{SubjectCode:} {Unique code identifying BHR participants}
 \item \code{TimepointCode:} {See BHR Codebook}
 \item \code{DaysAfterBaseline.Sleep:} {See BHR Codebook (DaysAfterBaseline)}
 \item \code{StatusDateTime.Sleep:} {See BHR Codebook (StatusDateTime)}
 \item \code{QID142:} {See BHR Codebook}
 \item \code{QID143:} {See BHR Codebook}
 \item \code{QID94:} {See BHR Codebook}
 \item \code{QID95.1:} {See BHR Codebook}
 \item \code{QID95.2:} {See BHR Codebook}
 \item \code{QID95.3:} {See BHR Codebook}
 \item \code{QID95.4:} {See BHR Codebook}
 \item \code{QID95.5:} {See BHR Codebook}
 \item \code{QID95.6:} {See BHR Codebook}
 \item \code{QID95.7:} {See BHR Codebook}
 \item \code{QID95.8:} {See BHR Codebook}
 \item \code{QID95.9:} {See BHR Codebook}
 \item \code{QID95.10:} {See BHR Codebook}
 \item \code{QID95.10.TEXT:} {See BHR Codebook}
 \item \code{QID96:} {See BHR Codebook}
 \item \code{QID97:} {See BHR Codebook}
 \item \code{QID98:} {See BHR Codebook}
 \item \code{QID145:} {See BHR Codebook}
 \item \code{QID146:} {See BHR Codebook}
 \item \code{QID101.1:} {See BHR Codebook}
 \item \code{QID101.2:} {See BHR Codebook}
 \item \code{QID101.3:} {See BHR Codebook}
 \item \code{QID101.4:} {See BHR Codebook}
 \item \code{QID101.5:} {See BHR Codebook}
 \item \code{QID101.5.TEXT:} {See BHR Codebook}
 \item \code{QID196:} {See BHR Codebook}
 \item \code{QID197:} {See BHR Codebook}
 \item \code{Sleep_Quality:} {PSQI Component 1. Calculated using BHR item QID96.}
 \item \code{Sleep_Latency:} {PSQI Component 2. Calculated using BHR items QID142 and QID95.1.}
 \item \code{Sleep_Duration:} {PSQI Component 3. Calculated using BHR item QID94.}
 \item \code{Sleep_Efficiency:} {PSQI Component 4. Calculated using BHR items QID94, QID141, and QID143.}
 \item \code{Sleep_Disturbance:} {PSQI Component 5. Calculated using BHR items QID95.2, QID95.3, QID95.4, QID95.5, QID95.6, QID95.7, QID95.8, QID95.9, and QID95.10.}
 \item \code{Sleep_Medication:} {PSQI Component 6. Calculated using BHR item QID97.}
 \item \code{Daytime_Fatigue:} {PSQI Component 7. Calculated using BHR items QID98 and QID145.}
 \item \code{PSQI_Total:} {Total score on the Pittsburgh Sleep Quality Index. Calculated by summing the seven components of the PSQI. Scores range from 0 to 21}
 }
}
\description{
This function cleanses data from the Brain Health Registry Sleep Module. The user is asked to input a raw dataframe from the BHR Sleep Module and to specify (1) the desired output variables and (2) the desired output time point(s) from which data was collected from each participant (i.e., first time point, last timepoint, etc.). The function performs basic data cleaning to remove duplicate observations (including obsertvations less than 30 days from the previous time point) and scores the Pittsburgh Sleep Quality Index (PSQI) using standard scoring algorithms .
}
